upd  field types

+++ application/account/models.py
@@ -1,9 +1,7 @@
 from datetime import datetime
 from typing import TYPE_CHECKING
 
-from sqlalchemy import Column, String, Boolean, Integer, TIMESTAMP, Date, ARRAY, select, func
-from sqlalchemy.dialects.postgresql.array import CONTAINS
-from sqlalchemy.ext.hybrid import hybrid_property
+from sqlalchemy import Column, String, Boolean, Integer, TIMESTAMP, Date, ARRAY, BigInteger, BIGINT
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from database.db import Base
@@ -18,7 +16,7 @@ if TYPE_CHECKING:
 class User(Base):
     __tablename__ = 'user'
 
-    id: Mapped[int] = mapped_column(Integer, unique=True, nullable=False, primary_key=True)
+    id: Mapped[BigInteger] = mapped_column(BIGINT, unique=True, nullable=False, primary_key=True)
     first_name: Mapped[str | None] = mapped_column(String(length=200))
     last_name: Mapped[str | None] = mapped_column(String(length=200))
     username: Mapped[str] = mapped_column(String(length=320), unique=True, index=True, nullable=False)
@@ -26,7 +24,7 @@ class User(Base):
     stack: ARRAY | None = Column(ARRAY(String), default=[])
     city: Mapped[str | None] = mapped_column(String(length=100))
     description: Mapped[str | None] = mapped_column(String(length=1000))
-    subscriptions: ARRAY | None = Column(ARRAY(Integer), default=[])
+    subscriptions: ARRAY | None = Column(ARRAY(BIGINT), default=[])
 
     registered_at: Mapped[Date] = mapped_column(TIMESTAMP, default=datetime.utcnow)
     hashed_password: Mapped[str] = mapped_column(String(length=1024), nullable=False)
